refactor main.go to break program up
